chore(dataplane): Move pipeline to pipeline_old
